[Techtask][EWS-1445]Release new format for temperature and other from glucometer

See merge request elta/elta-android!263

+++ data/src/main/java/com/elta/android/data/features/devices/dto/GlucometerEventDto.kt
@@ -5,7 +5,7 @@ import org.threeten.bp.ZonedDateTime
 data class GlucometerEventDto(
     val id: String,
     val date: ZonedDateTime?,
-    val temperature: Int?,
+    val temperature: Double?,
     val value: Double?,
     val glucometerSerialNumber: String?
 )
